Fixed checksum check by changing variable data type from uint8_t to uint16_t

+++ src/md_robot_node/com.cpp
@@ -55,9 +55,9 @@ int InitSerialComm(void)
     }
 }
 
-uint8_t CalCheckSum(uint8_t *pData, uint16_t length)
+uint16_t CalCheckSum(uint8_t *pData, uint16_t length)
 {
-    uint8_t sum;
+    uint16_t sum;
 
     sum = 0;
     for(int i = 0; i < length; i++) {
